+++ scanpipe/forms.py
@@ -158,13 +158,31 @@ class AddInputsForm(InputsBaseForm, forms.Form):
 
 class AddPipelineForm(PipelineBaseForm):
     pipeline = forms.ChoiceField(
-        choices=[
-            (name, pipeline_class.get_summary())
-            for name, pipeline_class in scanpipe_app.pipelines.items()
-        ],
         widget=forms.RadioSelect(),
         required=True,
     )
+
+    def __init__(self, project_runs=None, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        # The pipeline choices are determined based on the project context:
+        # 1. If no pipelines are assigned to the project:
+        #    Include all base (non-addon) pipelines.
+        # 2. If at least one pipeline already exists on the project:
+        #    Include all addon pipelines and the existing pipeline (useful for
+        #    potential re-runs in debug mode).
+        project_run_names = {run.pipeline_name for run in project_runs or []}
+
+        pipeline_choices = [
+            (name, pipeline_class.get_summary())
+            for name, pipeline_class in scanpipe_app.pipelines.items()
+            # no pipelines are assigned to the project
+            if (not project_runs and not pipeline_class.is_addon)
+            # at least one pipeline already exists on the project
+            or (project_runs and (name in project_run_names or pipeline_class.is_addon))
+        ]
+
+        self.fields["pipeline"].choices = pipeline_choices
 
     def save(self, project):
         self.handle_pipeline(project)